Create Plotting to see the relationship between Damper Positions and Vehicle Velocity.m
Here is MATLAB code for the plotting of the relationships to see the relationship between Damper Positions [Front Left (FL), Front Right (FR), Rear Left (RL), Rear Right (RR)] and Vehicle Velocity. This was done early in the project to visually see if there is some sorta relationship between the variables. From this, we can see what sort of Simple/Multi-linear Regression can be used to formulate shock-pod data as a function of velocity. This is used to find the aerodynamic center of balance so that it can be matched with the vehicle.